Implement update user api

+++ backend/src/entity/user.rs
@@ -3,7 +3,7 @@ use sqlx::FromRow;
 
 #[derive(Deserialize, Serialize, FromRow)]
 pub struct User {
-    pub id: u64,
+    pub id: i32,
     pub first_name: String,
     pub family_name: String,
     pub email: String,
@@ -12,7 +12,7 @@ pub struct User {
 
 impl User {
     pub fn new(
-        id: u64,
+        id: i32,
         first_name: String,
         family_name: String,
         email: String,
